[broken] attempt to support redrive_max_receive_count
This is a non-functional attempt to support redrive_max_receive_count.
What we want to do is create a condition in the template that decides
if we want to set a redrivePolicy for the primary queue, then use
Fn::If to conditionally set the policy, with references to the
secondary queue Arn and the given maxReceiveCount parameter.

It took a lot of hacks to even get this far.  Issues are:

 - the redrivePolicy parameter of the SQS resource is a weird
   JSON-string-in-JSON thing; we can't directly `Ref` things from
   within it, as far as I can tell, we have to instead use `Fn::Sub`
   to substitute the values into the JSON
 - the goformation.Sub() function doesn't support the two-argument
   version of Fn::Sub, which allows you to specify specific bindings.
   This is necessary because we need to use Fn::GetAtt to pull the
   secondary queue's ARN
 - goformation's Condition support is weird and rudimentary, and
   totally undocumented (like the rest of goformation)
 - when I actually try to render a template containing a Condition
   with a Fn::Equals call, the condition gets replaced with a literal
   `false`.  It looks like goformation is trying to evaluate the
   condition in place, and I really don't want it to do that.

It's worth asking why we're still using goformation?  It gives us:

 1. the ability to substitute in the queue name
 2. the ability to substitute in the tags
 3. nice types for everything

I'm starting to feel that 1. and 2. are completely doable without
goformation, and 3. is not nice enough to make up for goformation's
sharp edges.